requirements.txt
alembic==1.10.3
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
DateTime==5.1
exceptiongroup==1.1.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.29.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
zope.interface==6.0
